* fix VC8 warnings about signed/unsigned mismatch * do we really need HAVE_CONFIG_H for anything?
git-svn-id: http://miranda.googlecode.com/svn/trunk@7897 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/protocols/Yahoo/im.c
@@ -119,7 +119,7 @@ void ext_yahoo_got_im(int id, const char *me, const char *who, const char *msg, 
 			dbei.cbBlob = 2;
 			if (!CallService(MS_DB_EVENT_GET, (WPARAM)hEvent, (LPARAM)&dbei)) 
 				// got that event, if newer than ts then reset to current time
-				if (tm < dbei.timestamp) tm = time(NULL);
+				if ((DWORD)tm < dbei.timestamp) tm = (long)time(NULL);
 		}
 
 		pre.timestamp = tm;
